versions/utf8ru.pl: added processing of the symbol corresponding to the command \textdiv - recently just have encountered such a symbol in my documents and found a suitable UTF-8 equivalent.
tests/koitest.tex, tests/wintest.tex: accidentally figured out that the old
command \No is obsoleted in the current Texlive distribution, so I added it
as \newcommand to the .tex files so that they could be processed by
pdflatex without errors. And interchanged placements of the commands
\usepackage{fontenc} and \usepackage{babel} in preamble (fontenc should come
before babel), also just for pdflatex.