$WORDS_IN_INDEX configurable
1) If some $WORDS_IN_NAVIGATION_PANEL_TITLES are configured (4 by default),
and $SHOW_SECTION_NUMBERS (-show_section_numbers) is set, then one title
word less is displayed as one word of 4 is counted for the section number.
This is inconvenient. I changed it so that $SHOW_SECTION_NUMBERS, if set,
automatically adds one to the $WORDS_IN_NAVIGATION_PANEL_TITLES count, so
that the number of actual title words displayed remains independent of the
value of $SHOW_SECTION_NUMBERS.

2) I added $WORDS_IN_INDEX to l2hconf.pm, and to dot.latex2html-init in the
way analogous to $WORDS_IN_NAVIGATION_PANEL_TITLES (by default it is equal
to $WORDS_IN_NAVIGATION_PANEL_TITLES if not set explicitly). So, it is
configurable now as well.

3) The corresponding piece in styles/makeidx.perl was coded so that when
$SHOW_SECTION_NUMBERS (-show_section_numbers) was set, then only section
number was displayed in the Index, and no section name at all. There were no
possibility to display both. I changed this behavior: $WORDS_IN_INDEX
(configurable now) determines the number of words, $SHOW_SECTION_NUMBERS (if
set) adds section number, and if one needs to show only the number in Index
and no words, can either explicitly define $WORDS_IN_INDEX = 0, or use the
option -short_index.