LaTeX2HTML implementation of LaTeX package hyphenat
In the latex package, these commands allow hyphenation within the
words on either side of the dot, colon, slash and hyphen.  This
LaTeX2HTML package doesn't do anything to make hyphenation happen
within words in HTML, but it does allow the dot, colon slash or
hyphen in the text to come through.

There are still some commands from hyphenat.sty that are not
implemented here.

Fixes #80